refactor(contractspec): move framework funnels to runtime/internal/contractbuild — compiler-Hard upstream [#1038]
GREEN. NewFrameworkHTTP / NewEventDerivation move from kernel/contractspec to
runtime/internal/contractbuild. Because the package sits under runtime/internal/,
the Go compiler refuses imports from outside the runtime/ subtree, so business
code (cells/, examples/, cmd/, adapters/) physically cannot construct a framework
ContractSpec — a compiler-Hard upstream gate (same Medium→Hard internal/ upgrade
as #638 runtime/internal/authtest).

- NewFrameworkHTTP upstream open → Hard; FrameworkHTTPIDPrefix unexported.
- NewEventDerivation upstream Medium → Hard; single-file caller allowlist +
  drift guard retired (ceremonial once non-runtime callers are compiler-sealed
  and content is Validate()-closed).
- ContractSpec{} literal ban (downstream Hard) unchanged; contractbuild excluded
  as the sanctioned funnel home.
- callers switched: runtime/bootstrap, runtime/http/devtools, runtime/eventrouter.

ref: 范本 runtime/internal/authtest (#638)
