+++ runtime/bootstrap/health.go
@@ -18,9 +18,9 @@ import (
 	"net/http"
 
 	"github.com/ghbvf/gocell/kernel/cell"
-	"github.com/ghbvf/gocell/kernel/contractspec"
 	"github.com/ghbvf/gocell/runtime/auth"
 	"github.com/ghbvf/gocell/runtime/http/health"
+	"github.com/ghbvf/gocell/runtime/internal/contractbuild"
 )
 
 // Framework-internal ContractSpecs for health probe endpoints. These are not
@@ -28,12 +28,12 @@ import (
 // are registered by bootstrap itself rather than by a Cell RouteGroups
 // implementation. The "http.framework.health." spec ID prefix marks them as
 // runtime-internal, distinguishing them from cell-owned routes. The
-// contractspec.NewFrameworkHTTP funnel is the only legitimate ContractSpec
+// contractbuild.NewFrameworkHTTP funnel is the only legitimate ContractSpec
 // construction path under runtime/ per NO-MANUAL-CONTRACTSPEC-LITERAL-01.
 var (
-	specHealthLivez   = contractspec.NewFrameworkHTTP("http.framework.health.livez.v1", "GET", "/healthz")
-	specHealthReadyz  = contractspec.NewFrameworkHTTP("http.framework.health.readyz.v1", "GET", "/readyz")
-	specHealthMetrics = contractspec.NewFrameworkHTTP("http.framework.health.metrics.v1", "GET", "/metrics")
+	specHealthLivez   = contractbuild.NewFrameworkHTTP("http.framework.health.livez.v1", "GET", "/healthz")
+	specHealthReadyz  = contractbuild.NewFrameworkHTTP("http.framework.health.readyz.v1", "GET", "/readyz")
+	specHealthMetrics = contractbuild.NewFrameworkHTTP("http.framework.health.metrics.v1", "GET", "/metrics")
 )
 
 // HealthRouteGroupOption customizes the route groups returned by
