+++ runtime/eventrouter/contract_tracing_subscriber.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 
 	"github.com/ghbvf/gocell/kernel/cellvocab"
-	"github.com/ghbvf/gocell/kernel/contractspec"
 	"github.com/ghbvf/gocell/kernel/outbox"
 	"github.com/ghbvf/gocell/kernel/wrapper"
+	"github.com/ghbvf/gocell/runtime/internal/contractbuild"
 )
 
 // NewContractTracingSubscriber decorates an outbox.Subscriber so every
@@ -68,7 +68,7 @@ func (s *contractTracingSubscriber) Subscribe(
 	// legitimate runtime/ derivation path per NO-MANUAL-CONTRACTSPEC-LITERAL-01.
 	// Validate() runs inside the funnel, so a derived-spec error surfaces here
 	// with a wrapped context.
-	spec, err := contractspec.NewEventDerivation(
+	spec, err := contractbuild.NewEventDerivation(
 		sub.ContractID,
 		cellvocab.ContractKind(sub.ContractKind),
 		sub.ContractTransport,
