+++ tools/archtest/no_manual_contractspec_literal_test.go
@@ -1,56 +1,57 @@
 // INVARIANT: NO-MANUAL-CONTRACTSPEC-LITERAL-01
-//   - INVARIANT: CONTRACTSPEC-FRAMEWORK-BUILDERS-EXIST-01
 //
 // # NO-MANUAL-CONTRACTSPEC-LITERAL-01
 //
 // Invariant: contractspec.ContractSpec{…} composite literals and
 // contractspec.EventSpec(…) call expressions must only appear in:
 //   - generated/contracts/**/*_gen.go — business contract specs (codegen output)
-//   - kernel/contractspec/** itself  — type definition + typed funnels
-//     (NewFrameworkHTTP / NewEventDerivation, see framework.go)
+//   - kernel/contractspec/** itself  — the ContractSpec type definition
 //
 // Hand-written production code under cells/, examples/**/cells/, and runtime/
 // must NOT define ContractSpec literals. Framework-owned HTTP infra (health
 // probes, devtools catalog) and event-tracing derivations use the typed
-// funnels in kernel/contractspec/framework.go.
+// funnels NewFrameworkHTTP / NewEventDerivation in
+// runtime/internal/contractbuild (issue #1038 moved them there from
+// kernel/contractspec — see below).
 //
-// The two funnels enforce different invariants matched to their semantics:
-//
-//   - NewFrameworkHTTP — open caller (any runtime/ HTTP infra may construct
-//     a framework spec) but closed content (FrameworkHTTPIDPrefix panic at
-//     construction time).
-//   - NewEventDerivation — closed caller (single-file allowlist enforced by
-//     this archtest: only runtime/eventrouter/contract_tracing_subscriber.go)
-//     and closed content (Validate() embedded inside the funnel).
+// Funnel home + upstream gate (issue #1038): both funnels live in
+// runtime/internal/contractbuild. Because that package sits under
+// runtime/internal/, the Go compiler refuses imports from outside the runtime/
+// subtree (cells/, examples/, cmd/, adapters/, kernel/, tools/, tests/), so
+// business code physically cannot call them — a compiler-Hard upstream gate.
+// This is the same Medium→Hard upgrade issue #638 applied to
+// runtime/internal/authtest (ai-robust.md §Hard 范本目录 → "internal/ wrap 包").
+// The funnels construct ContractSpec{…} literals, so contractbuild is excluded
+// from this scan (the sanctioned funnel home, analogous to kernel/contractspec).
 //
 // Exclusions:
-//   - generated/contracts/**/*_gen.go  — the authoritative home for business contracts
-//   - tools/codegen/**/testdata/**     — codegen fixture files
-//   - **/fixtures/**                   — test fixture trees
-//   - kernel/contractspec/** itself    — defines ContractSpec and the typed funnels
-//   - *_test.go                        — test helpers may reference specs for assertions
+//   - generated/contracts/**/*_gen.go     — the authoritative home for business contracts
+//   - tools/codegen/**/testdata/**        — codegen fixture files
+//   - **/fixtures/**                      — test fixture trees
+//   - kernel/contractspec/** itself       — defines the ContractSpec type
+//   - runtime/internal/contractbuild/**   — the sanctioned runtime-side funnel home
+//   - *_test.go                           — test helpers may reference specs for assertions
 //
 // AI-robust:
-//   - Composite-literal ban: Hard — `contractspec.ContractSpec{…}` under
-//     cells/ + examples/ + runtime/ is unrepresentable (archtest fails CI),
-//     the typed funnels are the only surviving form.
-//   - NewEventDerivation content invariant: Hard — Validate() is embedded
-//     inside the funnel, so a malformed spec cannot be returned.
-//   - NewEventDerivation caller allowlist: Medium — enforced by path-string
-//     match against eventDerivationAllowedCaller. The drift guard
-//     TestEventDerivationAllowedCallerFileExists upgrades safety by failing
-//     CI when the constant goes stale, covering two modes: (a) the file
-//     disappears (Stat fails) and (b) the file remains but no longer invokes
-//     contractspec.NewEventDerivation (AST scan reports zero calls). The
-//     gate remains string-anchored per ai-robust.md taxonomy. Upgrade path
-//     to Hard would be a typed authority token only the eventrouter package
-//     can mint.
+//   - Composite-literal ban: Hard (downstream) — `contractspec.ContractSpec{…}`
+//     under cells/ + examples/ + runtime/ is unrepresentable (archtest fails
+//     CI), the typed funnels are the only surviving form.
+//   - NewFrameworkHTTP upstream: Hard — runtime/internal/ placement; non-runtime
+//     callers are a compile error. Content Hard via frameworkHTTPIDPrefix panic.
+//   - NewEventDerivation upstream: Hard — same runtime/internal/ placement.
+//     Content Hard via embedded Validate(). The former single-file caller
+//     allowlist (eventDerivationAllowedCaller) + its drift guard are RETIRED
+//     (#1038): once the compiler seals non-runtime callers and Validate() seals
+//     content, the "only eventrouter" rule guarded only a non-security tracing
+//     projection whose output must still pass Validate() — pure ceremony. See
+//     runtime/internal/contractbuild/doc.go for the single-source grading.
 //
 // Aligns with the "typed function call as Hard funnel for unbounded
 // operations" charter pattern (PANIC-REGISTERED-01 same path).
 //
 // ref: docs/plans/202605011500-029-master-roadmap.md K#PR4 W3 + G-04
-// ref: kernel/contractspec/spec.go construction-site catalog
+// ref: docs/reviews/202605181109-042-archtest-six-agent-audit.md §3b/§4 row 9
+// ref: runtime/internal/contractbuild/doc.go (#1038 funnel home + grading)
 package archtest
 
 import (
@@ -65,16 +66,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/ghbvf/gocell/kernel/cellvocab"
-	"github.com/ghbvf/gocell/kernel/contractspec"
 	"github.com/ghbvf/gocell/tools/archtest/internal/scanner"
 )
-
-// eventDerivationAllowedCaller is the single legitimate caller of
-// contractspec.NewEventDerivation. Any other production file invoking the
-// funnel fails NO-MANUAL-CONTRACTSPEC-LITERAL-01; renaming or relocating the
-// caller requires updating this constant in lockstep.
-const eventDerivationAllowedCaller = "runtime/eventrouter/contract_tracing_subscriber.go"
 
 // TestNO_MANUAL_CONTRACTSPEC_LITERAL_01 scans production .go files under
 // cells/, examples/*/cells/, and runtime/ for contractspec.ContractSpec{…}
@@ -100,25 +93,27 @@ func TestNO_MANUAL_CONTRACTSPEC_LITERAL_01(t *testing.T) {
 	}
 }
 
-// TestCONTRACTSPEC_FRAMEWORK_BUILDERS_EXIST_01 locks the typed funnel API.
-// If kernel/contractspec.NewFrameworkHTTP or NewEventDerivation are renamed,
-// removed, or their signatures change, this test fails to compile — signaling
-// that the Hard gate has lost its only legitimate runtime-side construction
-// path and must be revisited (either redirect callers to a new funnel or
-// update this archtest).
-func TestCONTRACTSPEC_FRAMEWORK_BUILDERS_EXIST_01(t *testing.T) {
-	t.Parallel()
-	_ = contractspec.NewFrameworkHTTP("http.framework.test.v1", "GET", "/test")
-	_, _ = contractspec.NewEventDerivation("event.test.v1", cellvocab.ContractEvent, "amqp", "test.topic")
-}
+// CONTRACTSPEC-FRAMEWORK-BUILDERS-EXIST-01 retired (#1038). The funnels moved
+// to runtime/internal/contractbuild, which tools/archtest cannot import (Go
+// internal/ rule — tools/ is outside the runtime/ subtree). The compile-time
+// existence lock now lives in the real runtime/ callers (runtime/bootstrap,
+// runtime/http/devtools, runtime/eventrouter) and contractbuild's own unit
+// test: a rename/removal/signature change of either funnel breaks those.
+
+// contractbuildFunnelDir is the sanctioned runtime-side ContractSpec funnel
+// home (runtime/internal/contractbuild). It legitimately constructs
+// ContractSpec{…} literals inside NewFrameworkHTTP / NewEventDerivation, so it
+// is excluded from this scan — analogous to the kernel/contractspec/** type
+// home being outside scope. See package doc (#1038).
+const contractbuildFunnelDir = "runtime/internal/contractbuild/"
 
 // collectContractSpecScanFiles returns production .go files to scan.
 // Scope: cells (top-level cells/ + examples/*/cells/) discovered via
 // findCellProductionGoFiles (metadata-driven), plus runtime/ via DirsScope
-// directory walk. kernel/contractspec itself owns the ContractSpec type
-// definition and the typed funnels (NewFrameworkHTTP / NewEventDerivation),
-// so it is intentionally outside this scope. *_gen.go files are excluded
-// from the unioned set.
+// directory walk. kernel/contractspec owns the ContractSpec type definition
+// and runtime/internal/contractbuild owns the typed funnels (NewFrameworkHTTP
+// / NewEventDerivation), so both are intentionally outside this scope.
+// *_gen.go files are excluded from the unioned set.
 func collectContractSpecScanFiles(t *testing.T, root string) []string {
 	t.Helper()
 	cellFiles, err := findCellProductionGoFiles(root)
@@ -133,6 +128,11 @@ func collectContractSpecScanFiles(t *testing.T, root string) []string {
 	out := make([]string, 0, len(cellFiles)+len(runtimeFiles))
 	for _, f := range slices.Concat(cellFiles, runtimeFiles) {
 		if strings.HasSuffix(f, "_gen.go") {
+			continue
+		}
+		// Exclude the sanctioned funnel home (runtime/internal/contractbuild).
+		if rel, relErr := filepath.Rel(root, f); relErr == nil &&
+			strings.HasPrefix(filepath.ToSlash(rel), contractbuildFunnelDir) {
 			continue
 		}
 		if _, dup := seen[f]; dup {
@@ -169,118 +169,11 @@ func TestNO_MANUAL_CONTRACTSPEC_LITERAL_01_NegativeFixture(t *testing.T) {
 	}
 }
 
-// TestNO_MANUAL_CONTRACTSPEC_LITERAL_01_NegativeFixture_NewEventDerivation
-// verifies that the scanner reports a violation when a file outside the
-// single-caller allowlist invokes contractspec.NewEventDerivation. A symmetric
-// positive case verifies the allowlist file (rel ==
-// eventDerivationAllowedCaller) is silent.
-func TestNO_MANUAL_CONTRACTSPEC_LITERAL_01_NegativeFixture_NewEventDerivation(t *testing.T) {
-	t.Parallel()
-	src := `package p
-import (
-	"github.com/ghbvf/gocell/kernel/cellvocab"
-	"github.com/ghbvf/gocell/kernel/contractspec"
-)
-func init() {
-	_, _ = contractspec.NewEventDerivation("event.bad.v1", cellvocab.ContractEvent, "amqp", "bad.topic")
-}
-`
-	tmp, err := os.CreateTemp(t.TempDir(), "newevent_test_*.go")
-	if err != nil {
-		t.Fatalf("create temp: %v", err)
-	}
-	if _, err := tmp.WriteString(src); err != nil {
-		t.Fatalf("write temp: %v", err)
-	}
-	if err := tmp.Close(); err != nil {
-		t.Fatalf("close temp: %v", err)
-	}
-
-	// Forbidden caller — simulated path outside the allowlist.
-	forbiddenRel := "runtime/fake/handler.go"
-	violations := scanForContractSpecLiterals(token.NewFileSet(), tmp.Name(), forbiddenRel)
-	if len(violations) == 0 {
-		t.Errorf("expected at least 1 violation for non-allowlist caller of NewEventDerivation, got 0")
-	}
-	for _, v := range violations {
-		if !strings.Contains(v, "NewEventDerivation") {
-			t.Errorf("violation message should mention NewEventDerivation: %q", v)
-		}
-		if !strings.Contains(v, eventDerivationAllowedCaller) {
-			t.Errorf("violation message should mention allowed caller %q: %q",
-				eventDerivationAllowedCaller, v)
-		}
-	}
-
-	// Allowed caller — simulated path matching the allowlist constant.
-	allowedViolations := scanForContractSpecLiterals(token.NewFileSet(), tmp.Name(), eventDerivationAllowedCaller)
-	if len(allowedViolations) != 0 {
-		t.Errorf("expected 0 violations for allowlist caller, got %d: %v",
-			len(allowedViolations), allowedViolations)
-	}
-}
-
-// TestEventDerivationAllowedCallerFileExists guards the caller-allowlist
-// constant against drift. Two failure modes are covered:
-//
-//  1. File missing — runtime/eventrouter/contract_tracing_subscriber.go is
-//     renamed or moved without updating eventDerivationAllowedCaller, so the
-//     allowlist matches nothing and every future production caller silently
-//     becomes allowed.
-//  2. Call moved within the package — the file still exists but no longer
-//     invokes contractspec.NewEventDerivation (e.g. the call was relocated to
-//     a sibling file in the same package). The allowlist path still matches
-//     for the original file but skips a check that no longer fires, weakening
-//     the gate without any test signal.
-//
-// Both modes fail fast so the constant must be updated in lockstep with any
-// reshuffling of the funnel call site.
-func TestEventDerivationAllowedCallerFileExists(t *testing.T) {
-	t.Parallel()
-	root := findModuleRoot(t)
-	target := filepath.Join(root, filepath.FromSlash(eventDerivationAllowedCaller))
-	info, err := os.Stat(target)
-	if err != nil {
-		t.Fatalf("eventDerivationAllowedCaller %q not found: %v", eventDerivationAllowedCaller, err)
-	}
-	if info.IsDir() {
-		t.Fatalf("eventDerivationAllowedCaller %q is a directory, want file", eventDerivationAllowedCaller)
-	}
-	// Verify the allowlisted file actually invokes
-	// contractspec.NewEventDerivation. Without this assertion, the allowlist
-	// could match a stale path whose call has migrated, silently weakening
-	// the gate (path-match short-circuits the violation report).
-	fset := token.NewFileSet()
-	f, err := parser.ParseFile(fset, target, nil, parser.SkipObjectResolution)
-	if err != nil {
-		t.Fatalf("parse %s: %v", eventDerivationAllowedCaller, err)
-	}
-	alias := contractspecLocalAlias(f)
-	if alias == "" {
-		t.Fatalf("%s no longer imports kernel/contractspec — drift; "+
-			"update eventDerivationAllowedCaller or restore the import",
-			eventDerivationAllowedCaller)
-	}
-	var found bool
-	scanner.EachInSubtree[ast.CallExpr](f, func(call *ast.CallExpr) {
-		sel, ok := call.Fun.(*ast.SelectorExpr)
-		if !ok {
-			return
-		}
-		ident, ok2 := sel.X.(*ast.Ident)
-		if !ok2 || ident.Name != alias {
-			return
-		}
-		if sel.Sel.Name == "NewEventDerivation" {
-			found = true
-		}
-	})
-	if !found {
-		t.Fatalf("%s no longer calls %s.NewEventDerivation — drift; "+
-			"update eventDerivationAllowedCaller to the new caller file or restore the call",
-			eventDerivationAllowedCaller, alias)
-	}
-}
+// NewEventDerivation caller-allowlist tests retired (#1038): the funnel moved
+// to runtime/internal/contractbuild, where the Go compiler seals non-runtime
+// callers; the former single-file ("only eventrouter") allowlist + its drift
+// guard guarded only a non-security tracing projection (Validate()-closed) and
+// are now pure ceremony. See package doc + runtime/internal/contractbuild/doc.go.
 
 // TestNO_MANUAL_CONTRACTSPEC_LITERAL_01_NegativeFixture_EventSpec verifies that
 // the scanner correctly identifies a contractspec.EventSpec() call expression.
@@ -325,15 +218,15 @@ func init() {
 // scanForContractSpecLiterals AST-scans f for:
 //  1. contractspec.ContractSpec{…} composite literals
 //  2. contractspec.EventSpec(…) call expressions
-//  3. contractspec.NewEventDerivation(…) call expressions outside the
-//     single-file allowlist (eventDerivationAllowedCaller)
 //
-// where "contractspec" is the local alias for kernel/contractspec.
+// where "contractspec" is the local alias for kernel/contractspec. Both forms
+// are forbidden globally outside generated/contracts/**/*_gen.go. The
+// NewFrameworkHTTP / NewEventDerivation funnels are NOT scanned here: they
+// moved to runtime/internal/contractbuild (#1038), where the Go compiler seals
+// non-runtime callers — a compiler-Hard upstream gate that needs no path-string
+// allowlist (see package doc).
 //
-// rel MUST be slash-separated (apply filepath.ToSlash at the caller). The
-// NewEventDerivation allowlist compares rel against eventDerivationAllowedCaller
-// by exact string match; an OS-native path on Windows would silently miss the
-// allowlist and report false violations.
+// rel MUST be slash-separated (apply filepath.ToSlash at the caller).
 func scanForContractSpecLiterals(fset *token.FileSet, path, rel string) []string {
 	data, err := os.ReadFile(filepath.Clean(path))
 	if err != nil {
@@ -366,9 +259,8 @@ func scanForContractSpecLiterals(fset *token.FileSet, path, rel string) []string
 			rel, pos.Line, alias,
 		))
 	})
-	// Match contractspec.EventSpec(…) and contractspec.NewEventDerivation(…)
-	// call expressions. EventSpec is forbidden globally outside generated/;
-	// NewEventDerivation is forbidden outside eventDerivationAllowedCaller.
+	// Match contractspec.EventSpec(…) call expressions — forbidden globally
+	// outside generated/contracts/**/*_gen.go.
 	scanner.EachInSubtree[ast.CallExpr](f, func(node *ast.CallExpr) {
 		sel, ok := node.Fun.(*ast.SelectorExpr)
 		if !ok {
@@ -378,22 +270,11 @@ func scanForContractSpecLiterals(fset *token.FileSet, path, rel string) []string
 		if !ok2 || ident.Name != alias {
 			return
 		}
-		switch sel.Sel.Name {
-		case "EventSpec":
+		if sel.Sel.Name == "EventSpec" {
 			pos := fset.Position(node.Pos())
 			violations = append(violations, fmt.Sprintf(
 				"%s:%d: manual %s.EventSpec() call — must be in generated/contracts/**/*_gen.go only",
 				rel, pos.Line, alias,
-			))
-		case "NewEventDerivation":
-			if rel == eventDerivationAllowedCaller {
-				return
-			}
-			pos := fset.Position(node.Pos())
-			violations = append(violations, fmt.Sprintf(
-				"%s:%d: %s.NewEventDerivation() call — only %s may invoke this funnel "+
-					"(derivation funnel is closed by caller-allowlist; see archtest doc)",
-				rel, pos.Line, alias, eventDerivationAllowedCaller,
 			))
 		}
 	})
